feat: folders restructured
