+++ stan-ditko/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/src/favicon.png" />
+    <link rel="icon" type="image/svg+xml" href="./src/assets/img/favicon.png"/>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Marvel.com | The Unofficial Site</title>
   </head>
