+++ stan-ditko/src/main.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import './index.css'
+import './assets/css/index.css'
 import App from './App'
 
 ReactDOM.render(
